Local merge

+++ src/main/java/matching/repository/MatchingRepository.java
@@ -7,12 +7,14 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import matching.bean.Matching;
+import matching.bean.MatchingDTO;
 import user.bean.DogsInfo;
+import user.bean.User;
 
 @Repository
-public interface MatchingRepository extends JpaRepository<DogsInfo, Integer> {
+public interface MatchingRepository extends JpaRepository<DogsInfo, Long> {
 	
-	public List<DogsInfo> findByUserId(int userId);
+	public List<DogsInfo> findByUserId(long userId);
 
 	public void save(Matching matchingBuilder);
